Ensure MAGIC_MAX_LENGTH is always aware of custom compression
Changes the way MAGIC is handled for the CustomCompressor by moving the
responsibility of setting the MAGIC property to the CompressionManager
constructor. This means that `CustomCompressor.MAGIC` will now have the
correct value once a CompressionManager has been created and we no
longer have to instantiate a CustomCompressor in order to determine its
magic bytes.

While this does move knowledge specific to the CustomCompressor into the
CompressionManager, there are two benefits:

  1. CustomCompressor no longer needs its own instance version of the
     class method `validate`.
  2. We no longer need to treat CustomCompressor as a special case when
     attempting to identify compression - we can just use the class
     method as with any other compressor.

We also change MAGIC_MAX_LENGTH from being a module variable to being
a property of the CustomCompressor. This means we can set it in the
CompressionManager constructor *after* we have set the MAGIC property
of CustomCompressor. This removes the need to update the value of
MAGIC_MAX_LENGTH via `sys.modules` when creating a CustomCompressor.

These changes fix a bug where files compressed with custom compression
could not be identified during `barman get-wal` operations due to
reading MAGIC_MAX_LENGTH bytes from a file before any CustomCompressor
had been created to update MAGIC_MAX_LENGTH. Specifically:

  1. `CompressionManager.identify` would read MAGIC_MAX_LENGTH bytes
     from a file.
  2. The method would then call the `validate` class method for each
     compressor class in the compression_registry, except for custom
     compression where it would create a CustomCompressor and call the
     `validate` instance method.
  3. If custom_compression_magic had a longer length than
     MAGIC_MAX_LENGTH then MAGIC_MAX_LENGTH would be updated with the
     length of custom_compression_magic.
  4. Unfortunately the bytes which are then passed to the validate class
     method have already been read back in step 1 using the old
     MAGIC_MAX_LENGTH value.
  5. Any custom_compression_magic longer than the original value of
     MAGIC_MAX_LENGTH (3 bytes) will therefore not be identified.

This bug could have been fixed by re-reading the bytes using the new
MAGIC_MAX_LENGTH value however this would have made the code even harder
to follow. This commit therefore applies a small refactoring to simplify
the code which deals with the compression at the cost of moving some
of the CustomCompressor knowledge into the CompressionManager.

Thanks to dsmoljanovic for finding and reporting this bug.

Closes #567.