Update some pg_start_backup references in comments
Replaces two mentions of pg_start_backup in the comments with
pg_backup_stop. This should have been updated in the commit which
added support for PostgreSQL 15.

Note there are some remaining references to pg_start_backup and
pg_stop_backup in the comments and code however the context of
those is that of an exclusive backup therefore the names are
appropriate.