Stop using `enum.Enum` in barman cloud
Replaces `enum.Enum` as a base class for UrlArgumentType and uses
`object` instead such that the "enum values" now become class variables.

This means we no longer rely on the enum34 library (the library which
contains the `Enum` backport`) on python2 systems. Relying on enum34
was bad for a number of reasons:

 1. We didn't declare it as a dependency and instead relied on other
    libraries to bring it in as a transtitive dep.
 2. The package is not available as an OS package on all supported
    systems which means users need to install it with pip.

Since the Enum was only acting as syntactic sugar it is preferable
to just avoid the dependency altogether.

This fixes a packaging bug where the enum34 library did not behave as
expected on some systems leading to import errors when running barman
cloud.

Note there are no test changes required because most of the barman
cloud unit tests are using UrlArgumentType indirectly and therefore
the class is adequately covered by those tests.

Closes #603.

Thanks to Ravi160492 for reporting this issue.