Open tarball for streaming when reading backup_label
Opens the backup tarball in mode "r|" so that TarFile treats the file
as a stream. This allows it to work with the file-like object produced
by the gzip decompressor on python 2 (in python 3 the mode was inferred
so it was not an issue, however specifying the mode does no harm so we
do it for both versions).