Remove backup_compression_* defaults from config
Removes the default values for backup_compression_format and
backup_compression_location from the config. This is because when
they are given default values commands like `barman diagnose` will
report those defaults, regardless of whether `backup_compression`
is set or not. This leads to providing information which is not
relevant to the user and could lead to confusion.

This commit therefore removes the defaults. These are not necessary
to get the default behaviour since pg_basebackup will also use those
same defaults if nothing is set.

There was one line of code which did rely on format being set - this
was used to determine whether the compression field should be set
in the backup info. For plain format compressed backups the field
should not be set, since the eventual on-disk format is the same as
an uncompressed backup. For everything else it should still be set.
This check is therefore updated from `== "tar"` to `!= "plain"`.