Remove default recovery staging path and add cli arg
Removes the default recovery staging path because there is no safe
assumption barman can make about where backup data should be staged
on recovery. Such a location needs to be large enough to store the
complete compressed backup and must meet the security needs of the
user. Barman can never know if a destination it chooses is suitable
therefore we require the user to specify it.

To make this a bit easier for users we add a cli arg which can be
used to set the recovery staging path, so they can just re-run their
recover command with the argument rather than edit the barman config.