Fetch config from recovery dir for mangling
Changes the way we fetch config files for mangling/analysis. Previously
we would use `tar tf` to check whether files were present in the tarball
and then extract them in a subsequent step, with both of these steps
happening after the tarball has been extracted into the final recovery
destination.

The problem with this is that we have to scan the tarball three times.
With sufficiently large base.tar.gz files this is going to add
significant time to the recovery process. The additional scans are
entirely avoidable given by this point in the process we have already
extracted the raw config files to the recovery destination where they
will eventually be overwritten with their mangled alternatives.

We therefore list the recovery destination directory to check for the
existence of config files and we rsync directly from there instead of
extracting from the tarball in the staging directory.

Once we're confident with this change we can remove the
`list_compressed_files` method and test since that will no longer be
needed.