Handle primary WAL switch for pg_basebackup backups
Adds an err_handler to the PostgresBackupExecutor command which checks
for the stderr line which indicates pg_basebackup is waiting for WALs.
When this line is received and we are backing up a standby PostgreSQL
instance, we manually perform a WAL switch on the primary so that the
pg_basebackup command can complete.

Relates to #596.