Add primary/standby checks
Adds new checks which only apply when `primary_conninfo` is set, indicating that
`conninfo` is expected to be pointing to a standby. In such cases Barman will
create a connection to both the primary and the standby - the main connection
is the standby and this is used for all postgres operations *except* switching
WALs, which is performed using the connection to the primary.

The new checks will:

1. Verify that the server reached via `conninfo` is a standby.
2. Verify that the server reached via `primary_conninfo` is not a standby.
3. Verify that both servers share the same system identifier.

These checks will fail if `conninfo` and `primary_conninfo` were misconfigured
such that they do not point to the standby and the primary respectively. The
checks will also fail in the event that the standby was promoted. In the event
that the standby was promoted the failing checks could be resolved by either
removing `primary_conninfo` and therefore backing up the primary as a regular
Barman server, or by updating the configuration such that `conninfo` points to
the new standby and `primary_conninfo` points to the new primary.

Relates to #596.