Update JSON output unit tests to use specific timezone
Updates JsonWriter tests to set a specific timezone for the functions
which output start or end epoch timestamps. This allows the tests to
verify that the epoch timestamps in the JSON output are correct w.r.t
the timezone.

Currently these tests will fail because of #618 - the use of
`strftime("%s")` means that the epoch timestamp will not be correct.