Treat existing backup_manifest file as an error condition
Updates the behaviour of `barman generate-manifest` such that if the
backup_manifest file already exists it is considered an error. Instead
of skipping the file creation and exiting with status 0, Barman will
now log the existence of the file at error level and exit with nonzero
status.

The reason for this change is that Barman cannot safely make the
assumption that the backup_manifest file is related to the backup. It
may be a manifest file from a previous backup, for example. It is
therefore up to the user to decide whether they can ignore the error
(in the case that they already know the generated manifest relates to
this particular backup) or remove the backup_manifest.

Closes #626.