[wip] Minimal PoC for backup via GCE snapshot
A new backup_method value "snapshot" is supported. If this is set then
the following config values should specify a GCE project and disk:

- snapshot_gce_project (global/server)
- snapshot_disk_name (server)
- snapshot_disk_zone (server)

The API call to take the snapshot is made in the barman process rather
than on the node on which the disk is mounted.

A new backup_info field snapshot_gce_project is added so we have a link
between the backup and the GCP project which stores the snapshot - i.e.
so that we can find it when we want to recover.

GCE is accessed via the google creds at the location pointed to by
GOOGLE_APPLICATION_CREDENTIALS. If a service account is used it needs
the following permissions:

- compute.disks.createSnapshot <-- for taking snapshots
- compute.snapshots.create <-- also for taking snapshots
- compute.globalOperations.get <-- for monitoring snapshot status
- compute.snapshots.list <-- for listing snapshots (currently not implemented)

The following additional things are needed (at least):

* Validate disk config in a check function:
  * Is it actually mounted on the server?
  * Is PGDATA there?
* Handle config files being in /etc and hence not necessarily in the snapshot.
* Make sure list-backup / show-backup / diagnose show relevant data for the snapshot.
* Check the health of the snapshots in list-backup / show-backup commands (maybe?).
* Optionally store snapshot to different location.
* Recover the backup.
* Delete the snapshot when a backup is deleted (not sure delete is currently
  an extensible thing in Barman).

Instead of having a SnapshotExecutor subclass for each cloud provider
we instead create a new interface which provides a take_snapshot
function. The SnapshotExecutor can then just be provided with whatever
CloudSnapshotInterface is needed.

Renames FsBackupExecutor to ExternalBackupExecutor and moves the
rsync-specific _start_backup_copy_message to the RsyncBackupExecutor
subclass.

The FsBackupExecutor duplicated code in SnapshotBackupExecutor is then
removed since it can not subclass ExternalBackupExecutor instead.

Note: This means SnapshotBackupExecutor now has:

  - ssh-specific checks
  - a notion of local mode

However both things will be necessary - ssh for checking mount points
and such, and local mode because undoubtably people will want to run
Barman on the DB node and do snapshots.