Set barman_lock_directory to /run/barman
Instead of letting this default to `barman_home` and recommending a
volatile location we instead specify `/run/barman` as the
`barman_lock_directory` value.

Combined with the use of systemd tmpfiles in the rpm/deb packages this
allows lock files to be stored on the volatile `/run` partition without
introducing issues around the removal of the `/run/barman` directory.

Part of #141.