[wip] list-backup support for cloud
Many issues:

* Does not report size or WAL size
* Does not determine WAL status
* Does not care about retention policy

These are actually not required for parity with the existing
barman-cloud-backup-list command but would clearly be useful.
For now perhaps we just don't report them instead of reporting them
as zero.

One thing which *is* preent in the barman cloud output but not the
barman output is the begin WAL value so perhaps that should be added.

Most importantly:

* Needs DRY