[wip] Take cloud backups with `barman backup`
Extends the CloudBackupManager.backup method to call the backup
executor and copies much of the old CloudBackupUploader logic into
the CloudBackupExecutor.

This does not support the "barman-cloud-backup as a hook script on
barman server" case. To do so would require having multiple
CloudBackupExecutor subclasses similar to the CloudBackupUploader
subclasses.

Quite how `barman backup` would be invoked as a hook script when it
reads the config from barman.conf for the server... I do not know.
Perhaps we need to add extra options to `barman backup` so that it
can be invoked in a one-off manner in a hook script. Or perhaps we
provide native support for the "upload backup to the cloud after
storing locally" case instead of requiring a hook script to be
used? This would probably work by chaining multiple backup executors.