[wip] Snapshot backup in Barman Cloud
Makes snapshot backups work in principle on Barman Cloud.

The implementation is a bit unsatisfactory because of the existing
inheritance hierarchy of the CloudBackupUploader class. There is a lot
of code around coordinating PostgreSQL backups in CloudBackupUploaderPostgres
which we'd like to re-use however there is also code and properties in that
class which we don't want at all. There are a few options:

1. Just subclass CloudBackupUploaderPostgres and make it do snapshots.
   This is quite the sleight of hand but it would allow code reuse and keep
   the snapshot code reasonably isolated, so not actually too horrible from
   a maintenance point of view.
2. Separate out the coordination and backup transfer logic in the CloudBackupUploader
   inheritance hierarchy so we have coordination classes which can then be composed
   with classes which actually do the backup mechanics. This results in code which,
   while technically more maintainable due to separation of concerns, is actually
   quite hard to follow.
3. Use mixins to add the PostgreSQL-specific functionality to the class which deals
   with taking the snapshot backup and the class which deals with uploading backups
   from a PostgreSQL server.

Alternatively burn it all and rewrite it, but that would be quite intrusive.

Options 1 and 3 result in code I could live with. This commit attempts option 3 and
extracts the code common to taking backups from PostgreSQL into a Mixin class. The
CloudBackupUploaderPostgres and CloudBackupSnapshot classes are both given the Mixin
which contains helpers and a _coordinate_backup method - this needs to be provided
with the backup_info, strategy and series of functions to carry out specific jobs
during the coordination of the backup.

Unfortunately the coordination logic can't know which arguments to provide to the
functions, only that they need to be called to make a thing happen. We therefore use
partials to create noarg versions of the functions which have already been given
everything they need to know.

It's a bit arbitrary in that it is literally just making the code reusable rather
than trying to create a meaningful abstraction, but at least it's honest.

Note that we do not re-use the upload controller for uploading the backup_label
so it ends up being an object rather than part of a 10KB tar file. We *could* have
reused the controller which would have given some advantages:

* Easier to make well factored code.
* Existing barman-cloud-restore command can fetch the backup_label.
* Consistent with existing barman cloud format in object store.
* If at some point in the future we need to store more stuff we can add
  it to the tarball.

Disadvantages:

* A lot of overhead just to store a few hundred bytes. You wouldn't do
  this is implementing a snapshot based thing from scratch.
* Looking like a regular barman cloud backup isn't necessarily a good
  thing - it could concievably be useful to look obviously different
  in the object store.