[wip] Snapshot backup in Barman Cloud
Makes snapshot backups work in principle on Barman Cloud. This is of
course a terrible implementation. Pretty much a copy/paste/edit of
bits of CloudBackupUploaderPostgres right into the cloud_backup module.

It works though, which is encouraging.

The next step is to find a way of sharing the code between the
uploading-a-barman-backup, uploading-a-pgdata, and taking-a-snapshot
backup approaches.