Handle PostgreSQL being unreachable during compression validation
Catches connection errors when determining the server version during
the validation of compression options in the PostgresBackupExecutor and
logs a warning instead of letting the exception break program execution.

This is required because the BackupExecutor is on the code path for most
Barman commands so if the exception is allowed to propagate then all
Barman commands for a server will fail when the actual PostgreSQL server
is unavailable. This is particularly absurd in the case of the `barman
recover` command, when in many cases the original PostgreSQL server will
indeed be unavailable.

In cases where the availability of PostgreSQL server should cause Barman
to fail we rely on the existing `Server.check_postgres` function which
is called during `barman check` and `barman backup`.

Closes #674.