Copy .partial WALs during recovery
Instead of attempting to copy .partial files during the same
`Rsync.from_file_list` call, we instead treat it as a separate copy.

Offers a number of improvements over the previous attempt:

1. `.partial` files are now also copied when WALs are not compressed.
2. The `.partial` file is not copied if it is not required due to
   the requested recovery time being met by existing WALs in the archive.

On the downside, the sonarqube cognitive complexity metrics are going
to hate this. There are a few opportunities for extracting bit of code
but both `get_required_xlog_files` and `_xlog_copy` are already quite
long and complex functions so it's unlikely we'll be able to make
sonarqube completely happy.