[sq] Only return the most recent .partial file if there are more than one
This "should never happen" however if `get_required_xlog_files` ever gets run
outside of the `barman recover` context it is possible that there are older
`.partial` files which haven't been dealt with by the archiver process yet.

In such cases we should make the same decision the archiver would and use the
latest file.