Warn about `.partial` files in the WAL archive during recovery
Updates _xlog_copy so that special handling is only used for `.partial`
files in the streaming directory. Any `.partial` files found in the
WAL archive are handled the same way Barman would previously handle
them, i.e. they are copied across to the recovery WAL destination
(decompressing if necessary) and *not* renamed. Because this is a
situation which requires some manual intervention a warning message is
now logged to tell the user about each `.partial` WAL found in the
archive.

The user must then determine whether the `.partial` WAL in the archive
is the result of a recovery from backup or a standby promotion and, if
so, can rename the file to omit its `.partial` extension.

There is a clear opportunity for Barman to handle this itself however
this is deliberately not tackled by this patch and will instead be
added at a later date when we improve the behaviour of Barman in a
clustered environment.

Relates to #393.