Batch delete object requests for GCP
Uses the `storage.Client.batch` context manager so that all delete
requests are executed in a single batch. In this first implementation
an exception is raised if the number of requests exceeds the maximum
allowed batch size - this will be addressed in the next commit where
batches up to the maximum allowed size will be created in the
CloudInterface class.

This is groundwork for #678.