[wip] Adding friendly names
An initial attempt which uses annotations to store the names.

The problem with this is that we have to do some extra work to get
that annotation into the various commands which display backup info.

It's also inefficient with the cloud because you have to read the
object value.

If we have to read the object value anyway we may as well just use
the backup.info file.

So... either create "lightweight" annotations and do the extra work
to add the name to where it needs to go, or just use backup.info.