[sq] add backup name to list_backups output
the plain text format is a bit messy if some backups have names and
others don't.

the idea is that ` - ` is the separator, so the first field is changed
from `server backup_id` to `server backup_id backup_name`. this is
likely less intrusive than trying to add on fields to the end after the
status field, which may or may not be present (and the retention target
field, which may or may not be present).

we could maybe have an option for including the backup name if it exists
which falls back to the backup ID if there is no backup name.

we *could* maybe take that policy everywhere.