[wip] support friendly names in barman cloud
This proved to be a little trickier than support in barman classic
because the cloud scripts are a little free-er in how and when they
deal with backup ids. The barman (classic) CLI will always reoslve them
before moving on, but the cloud scripts will often reference the original
config.backup_id. This can result in amusing scenarios such as a backup
getting deleted even though it has a keep annotation, because the backup
*name* was used to determine if an annotation exists, not the id.

So there is plenty of scope for this patch to screw things up. We should
therefore standardise barman cloud so the first thing they do is resolve
the backup_id - this means taking a trip through cloud storage *unless*
the backup_id looks like a backup id, in which case fine. It is better
to do it up front and always create a catalog (and therefore pass that
catalog to things that need it, since it will have cached everything -
this is necessary in the restore script).

TODO:

- create a parse_backup_id function for cloud, similar to the existing
  cli module
- it can just support names for now but should eventually support oldest
  latest etc
- make sure the catalog needed for that (since it will be a function on the
  catalog) gets passed to anything else that might otherwise create a
  catalog (I think this is just restore)