[wip] Add nocopy recovery executor
Turns out the logic we added for dealing with config files already on
the recovery host is useful for recovering from snapshots as well as
compressed tarballs.

We therefore move this into a new RemoteRecoveryExecutor which
TarballRecoveryExecutor subclasses *and* create a new
RemoteRecoveryExecutor subclass which just copies the backup_label.