[sq] tests for cloud parse_backup_id method
also fix the code so it throws an exception if there is no matching
backup.

the test is implemented using the in-memory cloud interface which was
previously only used for the keep test - this fixture needed to be
modified so that the list function lists paths according to the
delimiter, like the real cloud interfaces do (the keep tests didn't
care about this because they didn't rely on finding the "directory"
in which a backup.info file might live).