[sq] verify backup-cloud-backup-keep resolves backup names to IDs
This really doesn't look like much of a test but it's verifying that
barman keep cares about the result of parse_backup_id and not whatever
is passed in as the backup id. Though the logic of parse_backup_id is
not touched, the test will fail if keep command does not resolve the
backup name.