[sq] tentative cloud refactor
Here we move the in-line snapshot code into a CloudBackupSnapshot class
and make that and the CloudBackupUploader hierarchy inherit from CloudBackup.

This is not particularly earth shattering - the duplication in the `backup`
function of both CloudBackupSnapshot and CloudBackupUploaderPostgres still
exists.

However it is gearing up for a deeper refactor where we can either break
down the backup method into smaller reusable chunks or extract it into somthing
composable. Or perhaps a bit of both.