[sq] Use upload contorller to upload backup_label
Why store a few hundred bytes when you can store 10K instead...

...maybe this wasn't the best idea after all

However, I think we should run with this design. If it turns out we
want to avoid using a tarball in the end we can use the CloudUploadController
interface to give a simple uploader which just passes through to
cloud_interface.upload_fileobj.