[wip] Mostly refactored?
Can't work out if this is a bit too Keith Bank Account or not.

Breaks up the CloudBackupUploader hierarchy into CloudBackupCoordinator
and CloudBackupMechanism hierarchies.

The idea is that the coordination (which is different depending on
whether we're backing up PG directly or backing up an existing backup)
is separated from the backup execution (which is different depending
on whether it's streaming to the cloud from barman, streaming to the
cloud from postgres, or making snapshots).

In practice it's not amazing. On the plus side we get the reuse of
CloudBackupCoordinatorPostgres in the snapshot backups. But on the
downside:

1. It's more complicated - there are two separate inheritance hierarchies
   where the classes from each can be composed together (but not in every
   possible way).
2. It's harder to read - code which is logically still quite coupled is
   now in different places.
3. The abstraction doesn't really make sense - the coordinator still needs
   to know about a lot of mechanism related details like the location and
   id of an existing backup, the controller used to upload the backup
   (although this happens to be shared by everything now it may not always
   be the case).
4. It just reads weird.
5. The `backup_copy` method which is the interface between the coordinator
   and the mechanism doesn't abstract anything - we have to pass in details
   which are specific to the backup mechanism. We get away with it by having
   the snapshot mechanism ignore them. But if we're doing that... really what
   is the point.