Allow file for recovery config settings to be specified
Adds the option `--recovery-conf-filename` to `barman recover` which
allows the file used by Barman for writing recovery settings to be
specified. If used, this will override the default values of
`postgresql.auto.conf` (for PostgreSQL 12 and later) or `recovery.conf`
(for PostgreSQL 11 and earlier).

The motivation for this change is to allow Barman to write recovery
settings in scenarios where the default files are not available for
writing (for example when using `stolon`, the `postgresql.auto.conf`
file is a symlink to `/dev/null`).

Credit to mojtabash78 for testing this patch.

Relates to #771.