[refactor 1/5] Prepare CloudBackupUploader classes for new backup methods
In order to add new backup methods to barman cloud we need to separate
the coordination logic required to backup a live PostgreSQL instance
from the backup transport logic which copies the backup data to its
final location.

This was problematic because of the way we had added the ability to
upload existing backups when running as a hook script. The idea *was*:

    Uploading an existing backup from a Barman server reuses the
    existing backup transport logic but not the coordination logic.
    We should therefore create a common superclass which contains
    just the minimum backup transport logic shared by both types of
    backup since that is all that needs to be reused.

We therefore created CloudBackupUploader with the bare-bones PGDATA and
tablespace copying logic, then two subclasses: CloudBackupUploaderPostgres
(a specialisation for backing up a live postgres server) and
CloudBackupUploaderBarman (a specialisation for backing up an existing
Barman backup).

This was ok when the only backup transport was copying to cloud object
stores. In order to add snapshots we then want to reuse the coordination
logic in CloudBackupUploaderPostgres but we don't need any of the backup
transport logic. We therefore need to separate the coordination logic
and the backup transport logic somehow. This could be achieved by:

1. Extracting the backup logic into its own class hierarchy and then
   using composition to create an object which backups up *either* a
   PostgreSQL server or a Barman backup *and either* uses copying to
   cloud storage or a snapshot.
2. Extracting the coordination logic into a separate mixin class and
   using multiple inheritance to add the coordination required for a
   live PostgreSQL backup.

Neither option yielded code which seemed any easier to maintain than
just duplicating the common code between snapshot and object storage
backups. Furthermore the relationships between the classes weren't
really meaningful - everything could have really just been called
SmallestReusableCodeModuleForDoingThing.

This commit therefore refactors the way code is shared between the class
which backs up a live PostgreSQL server and the code which backs up an
existing backup on a Barman server. We make backing up a live PostgreSQL
server the main CloudBackupUploader class and then extend it with
the CloudBackupUploaderBarman class. This requires splitting the backup
copy into two functions so that CloudBackupUploaderBarman can use only
the code which copies the data files and not the config files (since the
data files will already be in the backup).

This makes sense because a CloudBackupUploaderBarman *is a*
specialisation of CloudBackupUploader where the backup has already been
taken (and therefore there is no coordination or config gathering to
do). The fact that the class does not need all the features of its
parent to do its job is neither here nor there. The key thing is the
relationship is meaningful to the humans who will have to maintain the
code.

This keeps all related code mostly in the same place which couldn't be
said for any of the attempts to solve this problem using composition
or multiple-inheritance.

In order to add snapshots we will extract the coordination logic into
the CloudBackup superclass so that a CloudBackupSnapshot class can
extend it. Currently CloudBackup is just a stub.