[refactor 2/5] Extract coordination logic from CloudBackupUploader
Following on from the previous commit, here we extract the coordination
logic and require subclasses to provide implementations of the various
things which need to happen.

Currently only those things which *will* be needed by snapshots are
extracted - code which is always shared is in-line. We might tidy this
up.