Add snapshot support to barman cloud
Implements snapshot backups as a feature of the barman-cloud-backup
command.

The feature is enabled by using the following arguments to
barman-cloud-backup:

    --snapshot-disks
    --snapshot-instance
    --snapshot-zone

Snapshot backup is only supported for the google-cloud-storage cloud
provider. This cloud provider requires the following additional
argument:

    --snapshot-gcp-project

Note: Although the backup consists of snapshots of persistent disks
stored by the cloud provider, the backup.info and backup_label are
not included in those snapshots and are instead uploaded to the
cloud object store specified by the destination_url.

Uploading backup.info to cloud storage allows the other barman cloud
commands to find and manipulate the backup metadata so commands like
barman-cloud-backup-list and barman-cloud-backup-keep will work with
snapshot backups.

This commit makes a small change to the barman-cloud-backup-delete
command so that snapshots are deleted at deletion time.

This commit also adds a new command, barman-cloud-backup-show, which
displays the snapshot metadata necessary to recover a backup from the
snapshots.

Relates to #620.