Catch ValueError when parsing backup ID
Catches any ValueError exceptions which occur when converting the
backup ID to a name and logs the result before exiting cleanly.

Previously we were just letting the exception travel to the top which
meant stack traces were dumped in the logs for no good reason. This
would also break the integration tests which will fail if any
exceptions or tracebacks are logged during the tests.

Relates to #690.