[refactor 1/2] Rename FsBackupExecutor for reuse
Renames FsBackupExecutor to ExternalBackupExecutor to reflect the fact
its key difference is that it is used for backups by external methods
using the low-level backup API rather than for pg_basebackup backups.

The _start_backup_copy method is moved into the RsyncBackupExecutor
class since it is specifc to the Rsync backup method.

This is groundwork for adding a new backup executor which will use
disk snapshots as its backup method.

Relates to #620.