[refactor 1/2] Rename FsBackupExecutor for reuse
Renames FsBackupExecutor to ExternalBackupExecutor to show that its
key difference is that it is used for backups by external methods
using the low-level backup API rather than for pg_basebackup backups.

The _start_backup_copy_message method is moved into the
RsyncBackupExecutor class since it is specific to the Rsync backup
method.

This is groundwork for adding a new backup executor which will use
cloud provider disk snapshots as its backup method.

Relates to #620.