[refactor 1/1] Create common CloudBackupDownloader superclass
Prepares CloudBackupDownloader for reuse by extracting the interface.

So that the backup metadata can be used to infer which downloader is
required, we change the constructor so that we accept a backup_info
already fetched from the catalog and no longer create the catalog
within the downloader.

We also move the check which fails if recovery destination dir has
content into the CloudBackupDownloaderObjectStore. This is because the
check only makes sense for the object store recovery where the backup
is streamed from the cloud. Other restore methods such as snapshots
actually require that the base backup is already restored into the
restore location, so should not fail on this check.

Relates to #620.