Add snapshot support to barman-cloud-restore
Adds a CloudBackupDownloaderSnapshot class which will perform a minimal
restore by downloading the backup_label to the recovery directory.

As with the SnapshotRecoveryExecutor it will also verify that disks
from the backup being restored have been cloned from the backup
snapshots, attached to the recovery instance and mounted at the correct
location with the expected mount options.

The CloudBackupDownloaderSnapshot is used automatically if the backup
being restored has snapshot metadata in the `snapshots_info` field of
its `backup.info`. The following arguments must be provided to the
barman-cloud-restore command:

    --snapshot-recovery-instance: The name of the compute instance.
    --snapshot-recovery-zone: The zone in which the compute instance
                              and disks are located.

Relates to #620.