[wip] Add snapshot backup method to Barman
A new backup_method value "snapshot" is supported. If this is set then
the following config values must be set:

  - snapshot_provider (global/server)
  - snapshot_gce_project (global/server)
  - snapshot_instance (server)
  - snapshot_disks (server)
  - snapshot_zone (server)

The API call to take the snapshot is made in the barman process rather
than on the node on which the disk is mounted.

A new backup_info field snapshot_gce_project is added so we have a link
between the backup and the GCP project which stores the snapshot - i.e.
so that we can find it when we want to recover. This will be replaced
in a subsequent commit a field containing all metadata required for a
recovery.

GCE is accessed via the google creds at the location pointed to by
GOOGLE_APPLICATION_CREDENTIALS. If a service account is used it needs
the following permissions:

  - compute.disks.createSnapshot <-- for taking snapshots
  - compute.snapshots.create <-- also for taking snapshots
  - compute.globalOperations.get <-- for monitoring snapshot status
  - compute.snapshots.list <-- for listing snapshots (*)
  - compute.disks.get <-- for gathering metadata (*)
  - compute.instances.get <-- for gathering metadata (*)
  - compute.snapshots.delete <-- for deleting snapshots (*)

Note: Permissions marked (*) are for features implemented in subsequent
commits.

Instead of having a SnapshotExecutor subclass for each cloud provider
we instead create a new interface which provides a take_snapshot
function. The SnapshotExecutor is then just provided with the
CloudSnapshotInterface for the specified cloud provider.

Relates to #620.